+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 + GH-88: Implement `groupIncreasing`, `groupDecreasing`, `groupNonIncreasing`, and `groupNonDecreasing` with both `Comparable` stream inputs or using an explicit `Comparator` to appropriately group elements in the input stream to lists in the output stream (thanks @nipafx)
 + Implement `ensureIncreasing`, `ensureDecreasing`, `ensureNonIncreasing` and `ensureNonDecreasing` with both `Comparable` stream inputs or using an explicit `Comparator` to ensure the given stream meets the criteria, or fail exceptionally otherwise
 + Implement `filterIncreasing`, `filterDecreasing`, `filterNonIncreasing` and `filterNonDecreasing` with both `Comparable` stream inputs or using an explicit `Comparator` to remove non-compliant elements from the input stream
++ Implement `repeat(n)` and `repeatInfinitely()` to repeatedly emit the input stream to the output stream
 
 ### 0.8.0 
 + Add support for `orElse()` and `orElseEmpty()` on size-based gatherers to provide a non-exceptional output stream
@@ -14,7 +15,7 @@
 + Implement `scanIndexed()` to perform a scan along with the index of each element
 + Implement `intersperse()` to put the given element between each element of the stream
 + Implement `cross(iterable)`, `cross(iterator)`, and `cross(stream)` to combine elements in the input stream with the given source of elements
-+ 
+
 ### 0.7.0
 + Use greedy integrators where possible (Fixes #57)
 + Add [JSpecify](https://jspecify.dev/) annotations for static analysis
@@ -38,7 +39,7 @@
 + Implement `reverse()`
 + Implement `maxBy(fn)` and `minBy(fn)`
 + Implement `exactSize(n)`
-+ 
+
 ### 0.4.0
 + Implement `suffle()` and `shuffle(RandomGenerator)`
 + Implement `filterWithIndex()`